Refactoring so Linear code is organized better.
